Minor code review comments
