+++ src/services/contract/utils.js
@@ -1,0 +1,14 @@
+const {utils} = require('ethers');
+
+function isValidAddress(address) {
+  try {
+    utils.getAddress(address);
+    return true;
+  } catch (e) {
+    return false;
+  }
+}
+
+module.exports = {
+  isValidAddress
+};
